feat: change withdrawals structure, add claim
