+++ packages/sdk/src/index.ts
@@ -1,13 +1,17 @@
 export { LidoSDK } from './sdk.js';
 export { type SDKError } from './common/utils/SDKError.js';
 export * from './common/decorators/index.js';
-export { StakeProps } from './staking/index.js';
 export {
-  RequestCallbackStages,
-  RequestStageCallback,
-  RequestProps,
-  RequestWithPermitProps,
+  TransactionCallbackStage,
+  type TransactionCallback,
+} from './core/index.js';
+export { type StakeProps } from './staking/index.js';
+export {
+  type ClaimRequestsProps,
+  type RequestProps,
+  type RequestWithPermitProps,
   ApproveCallbackStages,
-  ApproveStageCallback,
+  type ApproveStageCallback,
 } from './withdrawals/index.js';
 export { LIDO_CONTRACT_NAMES } from './common/constants.js';
+export { type WrapProps } from './wrap/index.js';
