+++ packages/sdk/src/withdrawals/claim/index.ts
@@ -1,0 +1,1 @@
+export { LidoSDKWithdrawalsClaim } from './claim.js';
