+++ packages/sdk/src/withdrawals/index.ts
@@ -1,9 +1,8 @@
 export { LidoSDKWithdrawals } from './withdrawals.js';
+export { ClaimRequestsProps } from './claim/types.js';
 export {
-  type RequestCallbackStages,
-  type RequestStageCallback,
   type RequestProps,
   type RequestWithPermitProps,
   type ApproveCallbackStages,
   type ApproveStageCallback,
-} from './types.js';
+} from './request/types.js';
