+++ packages/sdk/src/withdrawals/request/index.ts
@@ -1,0 +1,1 @@
+export { LidoSDKWithdrawalsRequest } from './request.js';
