+++ packages/sdk/src/wrap/index.ts
@@ -1,2 +1,2 @@
 export { LidoSDKWrap } from './wrap.js';
-export { type CommonWrapProps, type LidoSDKWrapProps } from './types.js';
+export { type WrapProps, type LidoSDKWrapProps } from './types.js';
